backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
faker>=20.0.0
google-genai>=1.0.0
supabase>=2.0.0
